current snapshot of work (wip, refactor pass next)
this is the full replacement snapshot for fusion right now. d5 26 ms, d6 200 ms, d7 4 seconds.

current remaining major issues: parity edge cases; i piece and t-spin edge cases; cleanup of over-optimization.

+++ crates/wasm/src/lib.rs
@@ -32,7 +32,7 @@ impl JsBoard {
         }
     }
 
-    pub fn from_rows(rows: &[u16]) -> Self {
+    pub fn from_rows(rows: &[u64]) -> Self {
         let mut board = Board::new();
         let limit = rows.len().min(Board::HEIGHT);
         for y in 0..limit {
@@ -64,8 +64,10 @@ impl JsBoard {
         self.inner.clear_lines()
     }
 
-    pub fn to_rows(&self) -> Vec<u16> {
-        (0..Board::HEIGHT).map(|y| self.inner.row(y)).collect()
+    pub fn to_rows(&self) -> Vec<u64> {
+        (0..Board::HEIGHT)
+            .map(|y| self.inner.row(y) as u64)
+            .collect()
     }
 }
